textDetection/requirements.txt
asgiref==3.8.1
Django==5.1
djangorestframework==3.15.2
numpy==2.1.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
pytesseract==0.3.13
sqlparse==0.5.1
typing_extensions==4.12.2
gunicorn